Adds theme support with theme manager
What
---

Added support for themes. A theme can override one or more of the base
templates. The active theme is set in the SiteConfiguration.
Additionally there is a basic theme manager and editor built into the
Django Admin.

Why
---

I want to be able to update basic styling / layout of my blog without a
deploy. I also want to be able to use this for more than one site, so
having themes is crucial.